web/backend/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
pydantic>=2.8.2
pydantic-settings>=2.3.0
sqlalchemy>=1.4.39
psycopg2>=2.9.10
loguru>=0.7.3
python-jose[cryptography] >=3.5.0
passlib[bcrypt] >=1.7.4
python-multipart >=0.0.20
python-dotenv >=1.0.1